Corregi el archivo Delete.jsx y pase el codigo al archivo queries.js
