+++ src/helpers/queries.js
@@ -28,3 +28,16 @@ export const crearProducto = async (productoNuevo) => {
 
 };
 
+
+export const borrarProducto = async (id) => {
+    try {
+        const respuesta = await fetch(APIProductos + "/" + id, {
+            method: "DELETE",
+        });
+        console.log(respuesta);
+        return respuesta;
+    } catch (error) {
+        console.log(error);
+    }
+};
+
